+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+* Add `--allCapabilities` to `create-key`
+
 ### Fixed
 * Fix testing bundle in CI for a new `staticx` version